Organized project files and moved extensions and view models to their own groups.
